+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - [#2093] AS3 Unnecessary use of fully qualified names for classes in same package
 - [#1678] Shapes - Miter clip join style
 - [#2094] AS3 do not show body trait variables as FQN
+- [#2094] AS3 Missing use namespace
 
 ## [19.0.0] - 2023-10-01
 ### Added